[IMP] Bye to crm_rma_lot_mass_return module: crm_rma_lot_mass_return module was joined with crm_claim_rma
