+++ crm_claim_rma/demo/sale_order_from_lot_mass.xml
