+++ crm_claim_rma/demo/stock_production_lot_from_lot_mass.xml
