+++ crm_claim_rma/demo/transfer_details_from_lot_mass.xml
