+++ crm_claim_rma/templates/assets.xml
@@ -3,7 +3,7 @@
     <data>
         <template id="assets_backend" name="search_product assets" inherit_id="web.assets_backend">
             <xpath expr="." position="inside">
-                <script type="text/javascript" src="/crm_rma_lot_mass_return/static/src/js/barcode_text.js"></script>
+                <script type="text/javascript" src="/crm_claim_rma/static/src/js/barcode_text.js"></script>
             </xpath>
         </template>
         <template id="layout" name="layout sound" inherit_id="web.layout">
@@ -11,8 +11,8 @@
                 <audio id="rma_alert">
                     <!-- TODO: maybe a module for this is right.
                      -->
-                    <source src="/crm_rma_lot_mass_return/static/src/sound/bip.ogg" type="audio/ogg"></source>
-                    <source src="/crm_rma_lot_mass_return/static/src/sound/bip.m4a" type="audio/m4a"></source>
+                    <source src="/crm_claim_rma/static/src/sound/bip.ogg" type="audio/ogg"></source>
+                    <source src="/crm_claim_rma/static/src/sound/bip.m4a" type="audio/m4a"></source>
                     Your browser does not support the audio element.
                 </audio>
             </xpath>
