+++ crm_claim_rma/tests/test_user_input.py
@@ -148,7 +148,7 @@ class TestUserInput(TransactionCase):
                     user_input[item['line']-1], safe_eval(item['value']))
 
     def test_02_invoice_search_validation(self):
-        invoice_id = self.env.ref('crm_rma_lot_mass_return.so_wizard_rma_1').\
+        invoice_id = self.env.ref('crm_claim_rma.so_wizard_rma_1').\
             invoice_ids
         self.assertEqual(len(invoice_id), 1, 'Expected only one invoice')
         wizard_id = self.wizard.with_context({
