+++ crm_claim_rma/views/crm_claim_from_lot_mass.xml
