+++ crm_claim_rma/wizards/__init__.py
@@ -26,3 +26,4 @@ from . import claim_make_picking
 from . import account_invoice_refund
 from . import res_config
 from . import claim_make_picking_from_picking
+from . import returned_lines_from_serial_from_lot_mass
