+++ crm_claim_rma/wizards/returned_lines_from_serial_from_lot_mass.xml
