--- crm_rma_lot_mass_return/i18n/es_VE.po
@@ -1,16 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-#	* crm_rma_lot_mass_return
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server 8.0\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2015-07-17 19:54+0000\n"
-"PO-Revision-Date: 2015-07-17 19:54+0000\n"
-"Last-Translator: <>\n"
-"Language-Team: \n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Plural-Forms: \n"
