+++ rma/__openerp__.py
@@ -30,7 +30,6 @@
         "crm_rma_location",
         "crm_rma_prodlot_invoice",
         "crm_rma_prodlot_supplier",
-        "crm_rma_lot_mass_return",
         "crm_claim_rma",
         "crm_claim_product_supplier",
         "crm_rma_claim_make_claim",
